+++ app/models/initiative.js
@@ -4,6 +4,7 @@ var Initiative = DS.Model.extend(Ember.Validations.Mixin, {
   title: DS.attr('string'),
   description: DS.attr('string'),
   isSupported: DS.attr('boolean'),
+  isClicked: DS.attr('boolean'), 
   issue: DS.belongsTo('issue'),
   suggestions: DS.hasMany('suggestion', {async: true}),