+++ tests/acceptance/making-a-suggestion-to-initiative-test.js
@@ -31,3 +31,23 @@ test('Successfully', function(){
     });     
   });
 });
+
+test('Manually typing url to make suggestion', function(){
+  visit('/initiatives/1/suggestions/new').then(function() {
+    expect( $("a:contains('Support this initiative')").length ).to.equal(1);
+    expect($("textarea").length).to.equal(0);
+    expect($(':submit').length).to.equal(0);
+    click( $("a:contains('Support this initiative')") ).then(function() {
+      expect(currentURL()).to.equal('/initiatives/1/suggestions/new');
+      expect( $("a:contains('Remove support for this initiative')").length ).to.equal(1);
+      expect( $("a:contains('Make a suggestion')").length ).to.equal(1);
+      click( $("a:contains('Make a suggestion')") ).then(function() {
+        expect($(':submit').attr('disabled')).to.equal('disabled');
+        fillIn('div.details textarea', "Make sure there's a doctor available for house calls");
+        click('form input[type=submit]').then(function() {
+          expect(find('.suggestion').first().text()).to.equal("Make sure there's a doctor available for house calls");
+        });
+      });
+    });
+  });
+});